requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
PyMySQL
Flask-Migrate
SQLAlchemy
Flask-Login
Flask-WTF
WTForms
Flask-Mail
PyJWT
Pillow
Flask-Babel
